move saml views into a module for further organization
